requirements.txt
# Dependências principais do aplicativo de upscale
#
# Instale torch/torchvision/torchaudio separadamente conforme seu hardware.
# Exemplos:
#   GPU (CUDA 12.1):
#     pip install torch==2.5.1+cu121 torchvision==0.20.1+cu121 torchaudio==2.5.1+cu121 --index-url https://download.pytorch.org/whl/cu121
#   CPU:
#     pip install torch==2.5.1 torchvision==0.20.1 torchaudio==2.5.1
#
# Depois:
#     pip install -r requirements.txt

realesrgan==0.3.0
basicsr==1.4.2
facexlib==0.3.0
numpy>=1.26
pillow>=10.0
opencv-python-headless>=4.10
requests>=2.32